RAML08Wrapper and related cleaned.
All non-atom-related removed or moved to the upper utils folder from atom/util.
Atom-related utils remain in atom/util folder.

Lots of unused dependencies removed.
