--- custom_typings/lrucache.d.ts
@@ -1,10 +1,0 @@
-declare module 'lrucache' {
-
-    interface CACHE{
-        set(k:string,v:any)
-        get(k:string):any
-    }
-    function lru (num: number): CACHE
-
-    export = lru
-}
