+++ package.json
@@ -46,28 +46,21 @@
     "atom-space-pen-views": "^2.0.5",
     "atom-ui-lib": "0.0.5",
     "basarat-text-buffer": "^4.1.5",
-    "base64url": "^1.0.4",
     "classnames": "^2.0.0",
     "client-oauth2": "^0.3.0",
     "emissary": "^1.3.3",
-    "esutils": "^2.0.2",
     "http-response-object": "^1.0.1",
     "invariant": "^2.0.0",
     "jquery": "^2.1.3",
-    "js-yaml": "^3.3.0",
     "jsonix": "^2.3.0",
     "know-your-http-well": "~0.2.0",
-    "lrucache": "^0.2.0",
     "marked": "^0.3.3",
     "mime": "^1.3.4",
     "mkdirp": "^0.5.0",
-    "node-libs-browser": "^0.5.2",
     "popsicle": "^3.2.2",
     "pretty-data": "^0.40.0",
     "react": "^0.13.2",
     "react-pure-render": "^1.0.1",
-    "request": "^2.53.0",
-    "traverse": "^0.6.6",
     "underscore": "^1.8.2",
     "w3c-schemas": "^1.2.1",
     "xhr2": "0.1.3",
@@ -75,20 +68,12 @@
     "xmlhttprequest": "^1.7.0",
     "xtend": "^4.0.1",
     "yaml-ast-parser": "0.0.26",
-    "z-schema": "^3.9.1",
     "raml-1-parser": "0.2.20",
     "raml-suggestions": "0.0.6"
   },
   "devDependencies": {
-    "chai": "^2.2.0",
-    "chai-as-promised": "^5.2.0",
-    "cheerio": "^0.19.0",
-    "typescript-compiler": "1.4.1-2",
     "coffee-loader": "^0.7.2",
     "coffee-script": "^1.9.2",
-    "dot": "^1.0.3",
-    "gaze": "^0.5.1",
-    "glob": "^5.0.15",
     "gulp": "^3.8.11",
     "gulp-compass": "^2.0.4",
     "gulp-connect": "^2.2.0",
@@ -96,19 +81,16 @@
     "gulp-rename": "^1.2.2",
     "gulp-sourcemaps": "^1.5.2",
     "gulp-typescript": "^2.11.0",
-    "highlights": "^1.3.0",
     "http-server": "^0.8.5",
     "json-loader": "^0.5.1",
     "mocha": "^2.2.1",
-    "nodemon": "^1.8.0",
     "q": "^1.0.0",
     "rimraf": "^2.3.2",
-    "serve": "^1.4.0",
     "source-map-loader": "^0.1.5",
     "string-replace-webpack-plugin": "0.0.2",
-    "typedoc": "^0.3.12",
     "typings": "^0.6.8",
     "webpack": "1.4.2",
+    "node-libs-browser": "^0.5.2",
     "typescript": "1.7.3",
     "dev-env-installer":"0.0.5",
     "run-sequence":"^1.0.0"
