+++ typings.json
@@ -23,10 +23,8 @@
     "htmlparser2": "github:DefinitelyTyped/DefinitelyTyped/htmlparser2/htmlparser2.d.ts#32f0a90dab26f70b206407ac24d4d622d0f0f408",
     "invariant": "file:custom_typings/invariant.d.ts",
     "jquery": "github:DefinitelyTyped/DefinitelyTyped/jquery/jquery.d.ts#01ce3ccf7f071514ff5057ef32a4550bf0b81dfe",
-    "js-yaml": "github:DefinitelyTyped/DefinitelyTyped/js-yaml/js-yaml.d.ts#52318243888c4aa02b3c0b5e68631577728ffcf8",
     "know-your-http-well": "file:custom_typings/know-your-http-well.d.ts",
     "loophole": "file:custom_typings/loophole.d.ts",
-    "lrucache": "file:custom_typings/lrucache.d.ts",
     "marked": "github:DefinitelyTyped/DefinitelyTyped/marked/marked.d.ts#d93cab373e6f15928b1d4475fe63a65691cdc468",
     "mime": "github:DefinitelyTyped/DefinitelyTyped/mime/mime.d.ts#7ca8add069411b3bde47fad7208706801cde8656",
     "mixto": "github:DefinitelyTyped/DefinitelyTyped/mixto/mixto.d.ts#be0b6b394f77a59e192ad7cfec18078706e44db5",
@@ -41,7 +39,6 @@
     "raml": "file:custom_typings/raml.d.ts",
     "react": "github:DefinitelyTyped/DefinitelyTyped/react/react.d.ts#be0b6b394f77a59e192ad7cfec18078706e44db5",
     "react-pure-render": "file:custom_typings/react-pure-render.d.ts",
-    "request": "github:DefinitelyTyped/DefinitelyTyped/request/request.d.ts#bba34156c1eaad788fa8996f7d30b54c4fce5ffe",
     "server-address": "file:custom_typings/server-address.d.ts",
     "source-map": "github:DefinitelyTyped/DefinitelyTyped/source-map/source-map.d.ts#be0b6b394f77a59e192ad7cfec18078706e44db5",
     "space-pen": "file:custom_typings/space-pen.d.ts",
